+++ src/Service/PudoService.php
@@ -11,6 +11,7 @@ use DPDProduct;
 use DPDPudo;
 use DPDShop;
 use Invertus\dpdBaltics\Config\Config;
+use Invertus\dpdBaltics\DTO\ShipmentData;
 use Invertus\dpdBaltics\Factory\ShopFactory;
 use Invertus\dpdBaltics\Provider\CurrentCountryProvider;
 use Invertus\dpdBaltics\Repository\ParcelShopRepository;
@@ -235,6 +236,32 @@ class PudoService
         $pudoOrder->city = $city;
         $pudoOrder->street = $street;
         $pudoOrder->save();
+    }
+
+    /**
+     * @param ShipmentData $shipmentData
+     * @param integer $idCart
+     *
+     * @return ShipmentData
+     * @throws \PrestaShopDatabaseException
+     * @throws \PrestaShopException
+     */
+    public function repopulatePudoDataInShipment($shipmentData, $idCart)
+    {
+        $pudoOrderId = $this->pudoRepository->getIdByCart($idCart);
+        $selectedPudo = $this->pudoRepository->getDPDPudo($pudoOrderId);
+
+        if (!$shipmentData->getSelectedPudoId() && !empty($selectedPudo->pudo_id)) {
+            $shipmentData->setSelectedPudoId($selectedPudo->pudo_id);
+        } elseif (!$shipmentData->getSelectedPudoIsoCode() && !empty($selectedPudo->country_code)) {
+            $shipmentData->setSelectedPudoIsoCode($selectedPudo->country_code);
+        } elseif (!$shipmentData->getCity() && !empty($selectedPudo->city)) {
+            $shipmentData->setCity($selectedPudo->city);
+        } elseif (!$shipmentData->getDpdStreet() && !empty($selectedPudo->street)) {
+            $shipmentData->setDpdStreet($selectedPudo->street);
+        }
+
+        return $shipmentData;
     }
 
     public function saveSelectedParcel($cartId, $city, $street, $countryCode, $idCarrier)